requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.10.4
asyncio==4.0.0
black==25.11.0
certifi==2025.10.5
click==8.3.0
coloredlogs==15.0.1
colorzero==2.0
dnspython==2.8.0
email_validator==2.2.0
gpiozero==2.0.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
humanfriendly==10.0
idna==3.11
loguru==0.7.2
mock==5.1.0
mypy==1.18.2
mypy_extensions==1.1.0
nats-py==2.7.2
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pydantic==2.8.2
pydantic-settings==2.12.0
pydantic_core==2.20.1
python-dotenv==1.0.1
pytokens==0.3.0
pytz==2024.1
RPi.GPIO==0.7.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.32
types-pytz==2025.2.0.20251108
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
